Project structure updates
- Mostly moving lots of files around.
- Added internal/ and public/ folders in routes/api and handlers/api to accommodate the two APIs Bema will have.
- Moved util files to new util folder.
- Files requiring other files now make use of process.cwd() to make file requiring easier.
- Routes were renamed and updated to use the correct HTTP methods, inspired by link below.
- JS forms make submit requests to updated API routes.

Inspiration behind some API design changes:
https://hackernoon.com/restful-api-design-step-by-step-guide-2f2c9f9fcdbf
