+++ index.js
@@ -2,13 +2,12 @@ require("dotenv").config();
 const PORT = process.env.PORT || 5000;
 const express = require("express");
 const bodyParser = require("body-parser");
-const urlencodedParser = bodyParser.urlencoded({extended: false});
 const cookieParser = require("cookie-parser");
-const db = require("./db");
+const db = require("./util/db");
 const errorHandler = require("./middleware/error");
 const log = require("./middleware/log");
 const isLoggedIn = require("./middleware/isLoggedIn");
-const apiRoutes = require("./routes/api");
+const apiRoutes = require("./routes/api/index");
 const pagesRoutes = require("./routes/pages");
 const authRoutes = require("./routes/auth");
 const app = express();
