+++ routes/pages.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const handlers = require("../handlers/pages");
+const handlers = require(process.cwd() + "/handlers/pages");
 
 router.get("/", handlers.home);
 router.get("/login", handlers.login);
