+++ util/functions.js
